feat: redesign month selector with centered title, gesture navigation, and cache
Replace the dropdown with a centered month title (full name), chevron
arrows, dot timeline, trackpad/touch swipe, and keyboard arrow support.
Add in-memory cache for instant month switching with per-mutation
invalidation. Document as Phase 10.

Made-with: Cursor

+++ src/utils.ts
@@ -17,7 +17,7 @@ export function formatYearMonth(ym: string): string {
   const year = parseInt(ym.slice(0, 4), 10);
   const d = new Date(year, month - 1);
   const locale = resolveLocale();
-  const monthStr = d.toLocaleDateString(locale, { month: "short" });
+  const monthStr = d.toLocaleDateString(locale, { month: "long" });
   return `${monthStr.charAt(0).toUpperCase()}${monthStr.slice(1)} ${year}`;
 }
